requirements.txt
asgiref==3.8.1
attrs==23.2.0
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
setuptools==70.3.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
